Reached above 70% average test coverage for imageList and questionList components of QA module
